[guppy] split up cargo.rs into several files
Going to make it easier to do upcoming work on summaries.
